tests: issue-217 Basic unit tests for route and service layers
