+++ backend/pyproject.toml
@@ -21,7 +21,7 @@ dependencies = [
     'pydantic==2.12.3',
     'matplotlib==3.5.0',
     'watchdog==4.0.0',
-    "flask==2.3.2",
+    "flask==3.1.0",
     "gunicorn==21.2.0"
 ]
 license = { file = "LICENSE" }
