+++ backend/tests/conftest.py
@@ -4,8 +4,13 @@
 
 import pytest
 import tempfile
+import typing
+
+import flask
+import flask.testing
 
 import spectre_server.core.config
+from spectre_server.__main__ import make_app
 
 
 @pytest.fixture
@@ -15,3 +20,19 @@ def spectre_config_paths():
         env = {"SPECTRE_DATA_DIR_PATH": tmpdir}
         paths = spectre_server.core.config.Paths(env)
         yield paths
+
+
+@pytest.fixture()
+def app() -> typing.Iterator[flask.Flask]:
+    app = make_app()
+    app.config.update(
+        {
+            "TESTING": True,
+        }
+    )
+    yield app
+
+
+@pytest.fixture()
+def client(app: flask.Flask) -> flask.testing.FlaskClient:
+    return app.test_client()
